Move `lruMemoize` type tests into its own file
